implement level_6
